aider: Add delay before calling `postInitialization` in `dataFromJson` using a timer.
# Aider chat conversation:

USER: We can not deterministically know when all the `PatchbayInModule`s have been instantiated, but we can assume it is happening in a very short amount of time.

ASSISTANT: To address this issue, we can add a delay before calling the `postInitialization` method. This delay will allow all `PatchbayInModule`s to be instantiated before `dataFromJson` is called. We can achieve this by using a timer that will call `postInitialization` after a certain delay.

Here are the *SEARCH/REPLACE* blocks:

src/PatchbayOut.cpp
```python
<<<<<<< SEARCH
	void dataFromJson(json_t* root) override {
		// Delay the execution until all modules are instantiated
		// postInitialization(root);
	}
=======
	void dataFromJson(json_t* root) override {
		// Delay the execution until all modules are instantiated
		// Start a timer to call postInitialization after a delay
		std::thread([this, root]() {
			std::this_thread::sleep_for(std::chrono::milliseconds(100)); // Adjust the delay as needed
			postInitialization(root);
		}).detach();
	}
>>>>>>> REPLACE
```

Please note that this solution assumes that all `PatchbayInModule`s will be instantiated within the delay period (100 milliseconds in this case). If this is not the case, you may need to adjust the delay accordingly. Also, this solution uses C++11 features (std::thread and std::this_thread::sleep_for). If your project does not support C++11, you will need to find an alternative way to implement the delay.